update for rename of project
