+++ pom.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.anchorage</groupId>
-    <artifactId>anchoragefx</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
+    <groupId>com.anchorfx</groupId>
+    <artifactId>anchorfx</artifactId>
+    <version>0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
     
     <properties>
